# openapi.yaml
openapi: 3.0.3
info:
  title: DPA Next API
  description: |
    DPA Next 智能研究助手 API
    
    提供项目生命周期管理、智能文档处理、认知分析等功能的RESTful API接口。
  version: 1.0.0
  contact:
    name: DPA开发团队
    email: support@dpa.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8001/api/v1
    description: 本地开发服务器
  - url: https://api.dpa.ai/v1
    description: 生产服务器
security:
  - bearerAuth: []
...
paths:
...
  /cognitive/analyze:
    post:
      tags:
        - cognitive
      summary: 文档深度分析
      operationId: analyzeDocument
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnalysisRequest'
      responses:
        '200':
          description: 分析结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
  /cognitive/chat:
    post:
      tags:
        - cognitive
      summary: 认知对话
      operationId: cognitiveChat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CognitiveChatRequest'
      responses:
        '200':
          description: 对话响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CognitiveChatResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    AnalysisRequest:
      type: object
      required:
        - document_id
      properties:
        document_id:
          type: string
          format: uuid
        analysis_type:
          type: string
          enum: [deep, quick, comparative]
        options:
          type: object
    AnalysisResponse:
      type: object
      properties:
        analysis_id:
          type: string
          format: uuid
        document_id:
          type: string
          format: uuid
        results:
          type: object
        confidence:
          type: number
        processing_time:
          type: number
    CognitiveChatRequest:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        project_id:
          type: string
          format: uuid
        context:
          type: object
    CognitiveChatResponse:
      type: object
      properties:
        response:
          type: string
        sources:
          type: array
          items:
            type: object
        confidence:
          type: number
        strategies_used:
          type: array
          items:
            type: string

    # Error schemas
...